Attempt basic attempt at adding HTML output.
The idea was to tag the source of each fragment, and present a good UI
for it, instead of litering with source location comments. Problem is
that `prettier` doesn't emit source maps, meaning that we are stuck with
ugly code we ourselves generate.

Solution is probably to, for each fragment, match the corresponding
pretty fragment, and then create new pretty fragment with their source
location intact.