Read source map when adding source locations.
The code may be generated with the location in the generator added as a
comment to each fragment of generated code. However, when running in
production, this led to the locations in our compiler versions being
used instead, which was less helpfull for debugging the generator. With
this commit, the generated file is searched for a source map
declaration, and if found, tries to resolve the fragments location to
our source files.

Note that the generated code now references source locations, while
those source locations aren't included in the created package.