Try to cast type better after validating body.
Idea ahere is that after running the jsonschema validator, typecast to
the expected type (instead of the previous any).

This is currently broken, since we at that point haven't handled string
format transformations yet, and they are instead all strings currently.
This technically works here, but it showed that `{ allOf: [anyOf: ...] }`
cases weren't implemented, leading to (correct) compilation errors in
some cases.