Fix import of Unlist symbol.
Old import in some cases led to the following error.

    export 'Unlist' (reexported as 'Unlist') was not found in '@todo-3.0/lib/unlist' (module has no exports)

This was due do the generated common file importing Unlist as a value,
but '@todo-3.0/lib/unilst' only exporting it as a type.

Simply adding `type` after `export` in the preamble solves the problem,
but we want to move away from magic symbols anyways, so this is a better
solution.